// front-end/src/app/modules/auth/models/authUser.configuration.model.ts
export interface AuthUserConfiguration {
    id?: number;
    authUserId :  number | null;
  
    allItemsSale:  boolean | null;
    countItemsSale :  number | null;
    printNoteSale:  boolean | null;
    
    allItemsShopping:  boolean | null;
    countItemsShopping :  number | null;
    printNoteShopping:  boolean | null;
}